Refactor file structure and update imports
