+++ day_13/ts/src/getCharacter.ts
