+++ day_13/ts/src/mathUtils.ts
