+++ day_13/ts/src/utils/gotDetails.ts
