+++ day_13/ts/src/utils/gotGreet.ts
