+++ day_13/ts/src/utils/gotUtils.ts
