+++ day_13/ts/tsconfig.json
@@ -7,6 +7,6 @@
     "skipLibCheck": true,
     "outDir": "./dist"
   },
-  "include": ["*.ts"],
+  "include": ["src/**/*.ts"],
   "exclude": ["node_modules"]
 }
