+++ day_13/ts/webpack.config.js
@@ -1,7 +1,7 @@
 const path = require("path");
 
 module.exports = {
-  entry: "./index.ts",
+  entry: "./src/index.ts",
   module: {
     rules: [
       {
